Fix printing non-existent bytes at end of file
This commit fixes an edge case where the contents of buf are printed,
even though those contents were never read from the file.

This issue was encountered while using the -A option with a match found
near the end of a file. Whatever contents were previously in buf were
printed, which is misleading since they didn't exist in the file.