Support External Pattern & Mask Files (#21)
Adds support for external pattern and mask files with the `-f` and `-m`
arguments, respectively. Both files are expected to be binary files
containing the bits to be matched. Files are read in byte order (any
endianness must be taken care of by the creator). At the moment, `-m` is
only available with `-f`, but that is not a strict limitation.

Increases the default buffer size to 1MB to increase throughput. For the
tested use case this reduced processing time by half.

Adds an `-r` option to enable recursion to match POSIX grep.

Fixes an off-by-one error when allocating space for the next path when
recursing.

Adds additional tests for using a pattern file as well as a pattern file
with a mask. Test failures now terminate early with a non-zero exit
code. `bgrep-test.sh` can now be run from any directory (such as the
project root).